fix bug in abstract writing
the curses library will error whenever you write to the bottom
right corner. In the rare cases that this happens, this
commit adds a simple check to catch that and move on.
Technically this could happen any time we write to a subwindow
but for now this is the only place I've seen it happen