add support for 2nd order ADS operators
Lets the new TOML config files specify some 2nd order operators
(e.g. citations, similar, etc.) which take in a (sub)query on their
own.
Also tweaks a couple details of the queries to better support the
new features of the new config files. A "keywords" term is not
required now (but is still often recommended).