refactor base controller to support new modes
this commit introduces the first structure of a more major change,
which is supporting 3 total modes, first the typical daily views
(lists and detailed), then an as-yet undefined library view (likely
based off of ADS library API) and finally a simple help page.

This commit adds a "base" view to support this by having the
basic `controller` open up to a simple menu with 4 buttons on it,
which recursively call the controller for that mode when selected.