add libraries functionality
This commit adds the next "mode", based on ADS libraries.
Uses the `ads.libraries` api and heavily subclasses a lot of
the code used for the daily mode, but based on a user's ADS
libraries instead of constructed queries.

This commit also cleans up and tweaks a number of other parts of
the code, so they work more generally and can be used by both
current modes more interchangeably.